Add files via upload
HR.xlsm is the excel file that one can use for payroll management in security industry. Adapted to CPF rules, details are in the Menu tab of the file.

HR Analytics is a Power BI analytics based on the data imported from HR.xlsm tables. 

Payslip Template helps to autogenerate payslips based on the PaymentData tab table that HR.xlsm has, Payslips (All) is what it is after the generation of the payslips.

Jan2016_... is a PDF file that contains the rules of CPF board that I have took reference from for the rules in terms of calculating CPF contributions. As I am not a HR Professional, my understanding of the document could be prone to errors.

The above system is done at my own time. As I am not a trained professional in software (I am a mechanical engineering major), I would gladly appreciate if any of you can give me guidance with regards to the code or even guide me to any web based programming stuff.